Split code into more components which is easy to manage
