+++ src/components/Sidebar.jsx
@@ -1,3 +1,5 @@
+import ProjectsList from './ProjectsList';
+
 function Sidebar({
   projects,
   openNewProject,
@@ -28,19 +30,7 @@ function Sidebar({
           </li>
         </menu>
 
-        <ul className="mt-8" data-testid="projects-list">
-          {projects.map((project) => (
-            <li key={project.id} className="my-1 mr-8">
-              <button
-                className={`w-full px-2 py-1 rounded-sm text-left text-ellipsis whitespace-nowrap overflow-hidden font-poppins text-lg tracking-wide ${project.selected ? 'bg-stone-800 text-stone-300' : 'text-stone-400'} hover:bg-stone-800 hover:text-stone-300`}
-                onClick={() => selectProject(project.id)}
-                data-testid={`select-${project.id}`}
-              >
-                {project.title}
-              </button>
-            </li>
-          ))}
-        </ul>
+        <ProjectsList projects={projects} selectProject={selectProject} />
       </div>
     </aside>
   );
